v2.5
Se desarrollo el controlador de indemizacion, aguinaldo, bono 14 y vacaciones.

Se completo el desarrollo del backend de un sistema de nominas.

+++ Sistema-Nominas/Models/Bono14.cs
@@ -9,9 +9,9 @@ public partial class Bono14
 
     public string? Empleado { get; set; }
 
-    public decimal? Meses { get; set; }
+    public int? Meses { get; set; }
 
-    public decimal? Dias { get; set; }
+    public int? Dias { get; set; }
 
     public decimal? Salario { get; set; }
